+++ src/php/StoryplayerInternals/SPv3/Defaults/defaultEnv.php
@@ -1,0 +1,5 @@
+<?php
+
+// This is the test environment setup/teardown script that we load *if*
+// the end-user doesn't use the -t/--target switch to load their own
+// test environment script.