+++ storyplayer/test-environments/vagrant-vbox-centos-6.7/devEnv.php
@@ -1,0 +1,5 @@
+<?php
+
+$env = null;
+
+return $env;